#158: use deleteReply to consolidate upgrade flow
